+++ Voxel2Pixel/Render/ArrayRenderer.cs
@@ -5,8 +5,8 @@ namespace Voxel2Pixel.Render
 	public class ArrayRenderer : TriangleRenderer
 	{
 		public byte[] Image { get; set; }
-		public int Width { get; set; }
-		public int Height => (Image.Length / Width) >> 2;
+		public ushort Width { get; set; }
+		public ushort Height => (ushort)((Image.Length / Width) >> 2);
 		#region IRectangleRenderer
 		public override void Rect(ushort x, ushort y, uint color, ushort sizeX = 1, ushort sizeY = 1) =>
 			Image.DrawRectangle(
